+++ src/plugin-main.h
@@ -7,7 +7,7 @@
 
 #include <obs-module.h>
 
-#define PLUGIN_VERSION "1.1.26"
+#define PLUGIN_VERSION "1.1.27"
 
 #ifdef __cplusplus
 extern "C" {